+++ src/main/java/org/reddot15/be_stockmanager/entity/Product.java
@@ -19,7 +19,7 @@ public class Product extends BaseMasterDataItem {
     String categoryName;
     Double importPrice;
     Double salePrice;
-    Double amount;
+    Integer amount;
     String earliestExpiry;
     Double vat;
